Change Socket::read back to &mut [u8]
Rust's std::io::Read trait reads into a &mut [u8] buffer, allowing it
to be a slice of a larger buffer, etc., not just a Vec. The previous
change I99dab579faadb0a493db17613f1117cc9b5ee981 changed it to a
Vec<u8>, but that broke the xi core module which was relying on it.

Change-Id: I5cf96d867b23db8c824806c6f5f53f1b325e9f0b